Migration/remove unused laps and metadata and p2p (#473)
* phpcs

* feat: add migration to remove all laps and their postmeta + p2p

* feat: add migration to rename laps_to_contacts to pg_relays_to_contacts

* chore: remove unused laps post type

* phpcs

* feat: add migration to remove unused metadata from relays

* fix: don't remove magic key metadata for now.

this will be saved for a magic link url refactoring

* chore: remove unneeded down from 0007 migration
