+++ post-type/loader.php
@@ -13,21 +13,13 @@ if ( ! class_exists( 'DT_Module_Base' ) ) {
  * Add any modules required or added for the post type
  */
 add_filter( 'dt_post_type_modules', function( $modules ){
-    $modules["relays_base"] = [
-        "name" => "Relays",
-        "enabled" => true,
-        "locked" => true,
-        "prerequisites" => [ 'contacts_base' ],
-        "post_type" => "pg_relays",
-        "description" => "Prayer Relays"
-    ];
-    $modules['laps_base'] = [
-        'name' => 'Laps',
+    $modules['relays_base'] = [
+        'name' => 'Relays',
         'enabled' => true,
         'locked' => true,
         'prerequisites' => [ 'contacts_base' ],
-        'post_type' => 'laps',
-        'description' => 'Prayer Laps'
+        'post_type' => 'pg_relays',
+        'description' => 'Prayer Relays'
     ];
     $modules['feedback_base'] = [
         'name' => 'Feedback',
@@ -41,9 +33,6 @@ add_filter( 'dt_post_type_modules', function( $modules ){
 
     return $modules;
 }, 20, 1 );
-
-require_once 'laps.php';
-Prayer_Global_Laps_Post_Type::instance();
 
 require_once 'relays.php';
 Prayer_Global_Relays_Post_Type::instance();
